main db changes + report frag actv addded

+++ app/src/main/java/com/example/rishabhtoys/TxnDetail.kt
@@ -2,6 +2,6 @@ package com.example.rishabhtoys
 
 data class TxnDetail(
      var date:String?,
-     var amount:Int?,
-     var balance:Int?) {
+     var amount:Float?,
+     var balance:Float?) {
 }
